+++ app/src/main/java/com/example/notes/NotesViewModel.kt
@@ -10,7 +10,7 @@ class NotesViewModel(private val repository: NotesRepository) : ViewModel(){
     val allNotes: LiveData<List<Notes>> = repository.allNotes.asLiveData()
     val favNotes: LiveData<List<Notes>> = repository.favNotes.asLiveData()
 
-    fun getNoteById(id: Long): LiveData<Notes> = repository.getNoteById(id).asLiveData()
+    fun getNoteById(id: Int): LiveData<Notes> = repository.getNoteById(id).asLiveData()
 
     fun insert(notes: Notes) = viewModelScope.launch {
         repository.insert(notes)
